feat: support cpu temperature detect and gpu monitor

+++ lib/src/field/power.rs
@@ -4,17 +4,17 @@ type Result<T> = std::result::Result<T, FieldError>;
 
 #[derive(Debug, Default, Clone, Decode, Encode)]
 pub struct Power {
-    value: f32, // Power consumption in Watts
+    value: u64, // Power consumption in mWatts
 }
 
 impl Power {
-    pub fn new(value: f32) -> Self {
+    pub fn new(value: u64) -> Self {
         Self { value }
     }
-    pub fn get_value(&self) -> f32 {
+    pub fn get_value(&self) -> u64 {
         self.value
     }
-    pub fn set_value(&mut self, value: f32) {
+    pub fn set_value(&mut self, value: u64) {
         self.value = value;
     }
 }
@@ -29,14 +29,14 @@ impl TryFrom<&[u8]> for Power {
 
 #[derive(Debug, Default, Clone, Decode, Encode)]
 pub struct TargetPower {
-    max: f32, // Power consumption in Watts
+    max: u64, // Power consumption in Watts
 }
 
 impl TargetPower {
-    pub fn new(value: f32) -> Self {
+    pub fn new(value: u64) -> Self {
         Self { max: value }
     }
-    pub fn get_max(&self) -> f32 {
+    pub fn get_max(&self) -> u64 {
         self.max
     }
 }
